refactor(core): give the deprecated names their own file
knip reads two exported names over one value as a duplicate, which is what a
deprecated alias is. Moving it out of package-config.ts keeps the check on for
the rest of the package and gives the next rename somewhere to put its alias.

The built declarations still carry both halves, checked the same way: a
consumer importing SystemConfig from the package gets the value and the type.
